separate sidebar components from layout, add folders for components with styles
