+++ src/Dropdown/index.jsx
@@ -1,0 +1,1 @@
+module.exports = {Dropdown: require('./Dropdown')}
