+++ src/SidebarLayout/index.jsx
@@ -1,0 +1,1 @@
+module.exports = {SidebarLayout: require('./SidebarLayout')}
