+++ src/SidebarLayout/sidebarlayout.less
@@ -1,0 +1,33 @@
+@import "~tfstyleguide/vars";
+
+.sidebar-layout-container {
+	display: flex;
+	min-height: ~"calc(100vh - 46px)";
+	align-items: stretch;
+	margin-top: -30px; // To account for nav margin
+	background-image: linear-gradient(
+		to right,
+		transparent 0%,
+		transparent 50%,
+		@white 50%,
+		@white 100%);
+	background-size: cover;
+}
+
+.sidebar-layout-sidebar {
+	width: (3 * @column-px) + (2 * @gutter-px);
+	margin-left: auto;
+	padding-left: @gutter-px;
+	flex-shrink: 0;
+	padding-top: 60px;
+}
+
+.sidebar-layout-main {
+	width: (10 * @column-px) + (9 * @gutter-px);
+	padding-right: @gutter-px;
+	margin-right: auto;
+	background-color: @white;
+	border-left: 1px solid @gray15;
+	padding-left: @column-px;
+	padding-top: 60px;
+}
