+++ src/SidebarMenuComponents/index.jsx
@@ -1,0 +1,3 @@
+module.exports = {
+  'MenuItem': require('./MenuItem'),
+  'SidebarMenu': require('./SidebarMenu')};
