+++ src/SidebarMenuComponents/sidebarmenucomponents.less
@@ -1,0 +1,62 @@
+@import "~tfstyleguide/vars";
+
+.menu-item {
+  .body-text-small;
+  display: block;
+  margin-bottom: 0;
+  cursor: pointer;
+  margin-left: -26px;
+
+  .icon-navigateright {
+    color: @blue;
+    transition: opacity .2s ease;
+    vertical-align: middle;
+    margin-right: 10px;
+    opacity: 0;
+  }
+
+  &.menu-item__active {
+    .proxima-bold;
+    position: relative;
+
+    .icon-navigateright {
+      opacity: 1;
+    }
+  }
+
+  a& {
+    color: @black;
+    &:hover {
+      color: @black;
+    }
+  }
+}
+
+.menu-item-parent {
+  color: @black;
+  &:hover {
+    color: @black;
+  }
+}
+
+.sidebar-menu {
+	margin-bottom: 45px;
+}
+
+.heading-small {
+	font-size: 14px;
+  color: @gray50;
+  text-transform: uppercase;
+  position: relative;
+  margin-bottom: 30px;
+
+  &::after {
+    content: '';
+    position: absolute;
+    width: 80px;
+    height: 3px;
+    background-color: @gray15;
+    bottom: -15px;
+    left: -3px;
+  }
+}
