+++ src/index.es6
@@ -1,6 +1,8 @@
 const {AppBar} = require('./AppBar');
 const {Dropdown} = require('./Dropdown');
 const {Gravatar} = require('./Gravatar');
-const {SidebarLayout, MenuItem} = require('./SidebarLayout');
+const {SidebarLayout} = require('./SidebarLayout');
+const {SidebarMenu, MenuItem} = require('./SidebarMenuComponents');
 
-module.exports = {AppBar, Dropdown, Gravatar, MenuItem, SidebarLayout}
+module.exports = {
+  AppBar, Dropdown, Gravatar, MenuItem, SidebarLayout, SidebarMenu}
