--- src/sidebarlayout.less
@@ -1,94 +1,0 @@
-@import "~tfstyleguide/vars";
-
-.sidebar-layout-container {
-	display: flex;
-	min-height: ~"calc(100vh - 46px)";
-	align-items: stretch;
-	margin-top: -30px; // To account for nav margin
-	background-image: linear-gradient(
-		to right,
-		transparent 0%,
-		transparent 50%,
-		@white 50%,
-		@white 100%);
-	background-size: cover;
-}
-
-.sidebar-layout-sidebar {
-	width: (3 * @column-px) + (2 * @gutter-px);
-	margin-left: auto;
-	padding-left: @gutter-px;
-	flex-shrink: 0;
-	padding-top: 60px;
-}
-
-.sidebar-layout-main {
-	width: (10 * @column-px) + (9 * @gutter-px);
-	padding-right: @gutter-px;
-	margin-right: auto;
-	background-color: @white;
-	border-left: 1px solid @gray15;
-	padding-left: @column-px;
-	padding-top: 60px;
-}
-
-.sidebar-menu {
-	margin-bottom: 45px;
-}
-
-.heading-small {
-	font-size: 14px;
-  color: @gray50;
-  text-transform: uppercase;
-  position: relative;
-
-  &::after {
-    content: '';
-    position: absolute;
-    width: 80px;
-    height: 3px;
-    background-color: @gray15;
-    bottom: -15px;
-    left: -3px;
-  }
-}
-
-.menu-item {
-  .body-text-small;
-  display: block;
-  margin-bottom: 0;
-  cursor: pointer;
-  margin-left: -26px;
-
-  .icon-navigateright {
-    color: @blue;
-    transition: opacity .2s ease;
-    vertical-align: middle;
-    margin-right: 10px;
-    opacity: 0;
-  }
-
-  &.menu-item__active {
-    .proxima-bold;
-    position: relative;
-
-    .icon-navigateright {
-      opacity: 1;
-    }
-  }
-
-  a& {
-    color: @black;
-    &:hover {
-      color: @black;
-    }
-  }
-}
-
-.menu-item-parent {
-  color: @black;
-  &:hover {
-    color: @black;
-  }
-}
-
